Main

+++ pkg/util/pagination/pagination.go
@@ -1,5 +1,5 @@
 package pagination
 
-func GetPageOffset(pageNum, pageSize int64) int64 {
+func GetPageOffset(pageNum, pageSize uint32) uint32 {
 	return (pageNum - 1) * pageSize
 }
